+++ client/app/main/editor.js
